move snarkfees to separate component file
